+++ image/smoke-tests/terraform/main.tf
@@ -49,9 +49,8 @@ module "proxy" {
   REGION  = var.region
   ZONE    = var.zone
 
-  NETWORK            = google_compute_network.this.name
-  SUBNETWORK         = google_compute_subnetwork.this.name
-  SUBNETWORK_PROJECT = google_compute_subnetwork.this.project
+  NETWORK            = google_compute_network.this.id
+  SUBNETWORK         = google_compute_subnetwork.this.id
 
   AUTO_CREATE_FIREWALL_RULES = false
   TRAFFIC_DISTRIBUTION_MODE  = "dns_round_robin"